+++ fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/RuleSet.java
@@ -107,6 +107,7 @@ import com.starrocks.sql.optimizer.rule.transformation.PushDownPredicateWindowRu
 import com.starrocks.sql.optimizer.rule.transformation.PushDownProjectToCTEAnchorRule;
 import com.starrocks.sql.optimizer.rule.transformation.QuantifiedApply2JoinRule;
 import com.starrocks.sql.optimizer.rule.transformation.QuantifiedApply2OuterJoinRule;
+import com.starrocks.sql.optimizer.rule.transformation.RemoteScanMinMaxConjunctsRule;
 import com.starrocks.sql.optimizer.rule.transformation.RemoteScanPartitionPruneRule;
 import com.starrocks.sql.optimizer.rule.transformation.RewriteBitmapCountDistinctRule;
 import com.starrocks.sql.optimizer.rule.transformation.RewriteDuplicateAggregateFnRule;
@@ -245,7 +246,10 @@ public class RuleSet {
                 PushDownPredicateToExternalTableScanRule.MYSQL_SCAN,
                 PushDownPredicateToExternalTableScanRule.JDBC_SCAN,
                 new MergeTwoFiltersRule(),
-                new PushDownPredicateCTEConsumeRule()
+                new PushDownPredicateCTEConsumeRule(),
+                RemoteScanMinMaxConjunctsRule.HIVE_SCAN,
+                RemoteScanMinMaxConjunctsRule.HUDI_SCAN,
+                RemoteScanMinMaxConjunctsRule.ICEBERG_SCAN
         ));
 
         rewriteRules.put(RuleSetType.PUSH_DOWN_SUBQUERY, ImmutableList.of(
